Adding setup.py and some PEP8 style
